Major Update

+++ database/migrations/2019_07_01_083840_add_font_branding_to_logodesign_orders.php
@@ -29,8 +29,11 @@ class AddFontBrandingToLogodesignOrders extends Migration
     {
         Schema::table('logodesign_orders', function (Blueprint $table) {
             //
-            $table->dropColumn('font');
-            $table->dropColumn('branding');
+            if(Schema::hasColumn('font', 'branding')) {
+                $table->dropColumn('font');
+                $table->dropColumn('branding');
+            }
+           
         });
     }
 }
